API: Update Requirements to use Injector and Config
