+++ _config/requirements.yml
@@ -1,0 +1,5 @@
+---
+Name: requirements
+---
+Injector:
+  Requirements_Backend: RequirementsHandler
