+++ docs/en/04_Changelogs/4.0.0.md
@@ -1,0 +1,30 @@
+# 4.0.0 (unreleased)
+
+## Overview
+
+### Framework
+
+ * `SapphireTest` now nests `Injector` and `Config` before each test and unnests them after
+
+### Other
+
+ * The `Requirements` class has been updated to make use of the `Injector` and `Config` APIs, the
+   following methods/settings are marked as deprecated and will be removed in 5.0:
+	* Requirements_Backend configuration:
+		- `Requirements_Backend` has been renamed to `RequirementsHandler`, with `Requirements_Backend` becoming an interface
+		- `Requirements::set_backend()` has been replaced with `Injector` configuration, see the [requirements developer guide](/developer_guides/templates/requirements#extending--customising) for more information
+	* Renamed methods:
+		- `Requirements::customScript()` has been replaced with `Requirements::custom_script()`
+		- `Requirements::customCSS()` has been replaced with `Requirements::custom_css()`
+		- `Requirements::insertHeadTags()` has been replaced with `Requirements::insert_head_tags()`
+		- `Requirements::javascriptTemplate()` has been replaced with `Requirements::javascript_template()`
+		- `Requirements::themedCSS()` has been replaced with `Requirements::themed_css()`
+		- `Requirements::includeInHTML()` has been replaced with `Requirements::include_in_html()`
+	* Configuration:
+		- `Requirements::set_combined_files_enabled()` has been replaced with the `Requirements.combined_files_enabled` config setting
+		- `Requirements::set_combined_files_folder()` has been replaced with the `Requirements.combined_files_folder` config setting
+		- `Requirements::set_suffix_requirements()` has been replaced with the `Requirements.suffix_requirements` config setting
+		- `Requirements::set_write_js_to_body()` has been replaced with the `Requirements.write_js_to_body` config setting
+		- `Requirements::set_force_js_to_bottom()` has been replaced with the `Requirements.force_js_to_bottom` config setting
+		- Setting `Requirements::backend()->write_header_comment` has been replaced with the `Requirements.write_header_comment` config setting
+		- Setting `Requirements::backend()->combine_js_with_jsmin` has been replaced with the `Requirements.combine_js_with_jsmin` config setting
